merges WorkshopSeats into Workshops. This prepares for a more aligned data model so that the inputs to the distribution algorithm are Topics: topic id -> topic attributes like category Workshops: workshop id -> workshop attributes like topic, timeslot, seats and Students workshop selections.
As consequence, the BiMap between WorkshopId and TopicTimeslot is given up. To compensate for this strong guarantee, the verification now contains a check that there is a 1:1 relationship.

Actually, while the WorkshopSeats type is removed from the HCD model, it remains as datatype inside the fullcombinatoric Algorithm, as it is there useful for the recursion and return value.

As a further consequence, there is no need anymore to have the combined type TopicTimeslot using a case class. TopicTimeslot used to be a combined type as it served as unique entity in the BiMap. Now that the BiMap is replaced by a verification, the TopicId and Timeslot can also now be broken apart, as otherwise there is an imbalance with adding more attributes like Seats to the attribute tuple.