allows the selection of a distribution algorithm on the command line, with a default.
As consequence, the type of the distribution algorithm must be fixed, i.e. have no type parameter. This makes sense as from the main method's point of view the algorithm shall always behave the same, either have side effects, but better return a defined type of result, so that the main method can continue with it, here currently calculating a metric and just printing it out.

Therefore, we give up on having a type A for additional return types of a DistributionAlgorithm. In consequence, we remove the extra result types from the DistributionAlgorithm and helper methods like Verification.

But as for the fullcombinatoric algorithm the extra returned values like the metric as calculated by the algorithm and the left-over free seats give valuable insight into the algorithm which also can be tested against, we keep them in the algorithm implementation and only remove them on the public DistributionAlgorithm method. We can however remove the extra tupling of the internal metric and left free seats, which was only introduced as we needed a single value for the extra type A. Removing the tupling simplifies the test cases.

Also, we now cannot wrap a "withMetric" around a distribution algorithm anymore, but it must now be a standalone operation.

We also need an enumeration of the available algorithms, thus we introduce enumeratum. The cool thing on scopt is that we can create a enumeration type such that we can use it type-safe in the command line parser.