adds test for the fact that each student can only have 1 workshop per timeslot. This triggers that the timeslot in taken into use in the recursion.
Originally we could use the BiMap SelectedTopics, later the Set[Topics] for each student, but both have no guaranteed order. This was fine as long as we did not pay attention on the timeslots of the workshops.

Now that we start to consider the timeslots, it makes a difference which workshop/timeslot is chosen for which topic. For the unit tests we need a reliable order in which the findWorkshopId searches for the next workshop which fits to a topic, therefore we need to also order the selected topic, here by selection priority. This also changes the expectation on existing tests.

Together with the rigid simple in-order assignment of students to workshops this means that there are situations that no distribution can be found during the recursion, even if they could be found with a more elaborate search. We have a test case for this, but it currently only records the fact that in this specific case the student does not get a replacement assignment.

The functionality of correct replacement assignments is more complicated and needs to be formulated during the next commit(s).